Various tidying up jobs. V.150.1 has moved forward, but is still a work in progress. The various modules using Goertzel filters were not using a harmonised approach to setting thresholds and measuring power levels. This has been improved.

+++ src/spandsp/private/sprt.h
@@ -107,15 +107,15 @@ struct sprt_state_s
 
     struct
     {
-        int subsession_id;
-        int payload_type;
+        uint8_t subsession_id;
+        uint8_t payload_type;
 
         sprt_chan_t chan[SPRT_CHANNELS];
     } rx;
     struct
     {
-        int subsession_id;
-        int payload_type;
+        uint8_t subsession_id;
+        uint8_t payload_type;
 
         sprt_chan_t chan[SPRT_CHANNELS];
